requirements.txt
# Framework Web
Flask==3.0.0
Werkzeug==3.0.1

# Machine Learning
tensorflow==2.15.0
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-learn==1.3.2

# Data Processing
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

# Database
psycopg2-binary==2.9.9

# Security
python-dotenv==1.0.0
Flask-WTF==1.2.1

# Production Server
gunicorn==21.2.0

# Development (optionnel)
# pytest==7.4.3
# black==23.12.1
# flake8==6.1.0